+++ sources/test_code/hd44780u_test.c
@@ -200,7 +200,7 @@ void hd44780u_test_app()
 
     // Control flags 
     uint8_t arg_convert = 0; 
-    uint16_t set_get_status = 0; 
+    uint32_t set_get_status = 0; 
     uint8_t cmd_index = 0; 
     uint8_t state = 0;
